requirements.txt
google-genai
pydantic
pytest
langchain
langgraph
langchain-google-genai
